Suggestions from @martinthomson
